+++ typescript/rebalancer-sim/test/utils/simulation-helpers.ts
@@ -1,4 +1,3 @@
-import { expect } from 'chai';
 import { ethers } from 'ethers';
 import * as fs from 'fs';
 import * as path from 'path';