+++ typescript/sdk/src/token/EvmWarpModule.ts
@@ -62,6 +62,7 @@ import { ChainName, ChainNameOrId } from '../types.js';
 import { extractIsmAndHookFactoryAddresses } from '../utils/ism.js';
 
 import { EvmWarpRouteReader } from './EvmWarpRouteReader.js';
+import { EvmXERC20Module } from './EvmXERC20Module.js';
 import { DeployableTokenType, TokenType } from './config.js';
 import { resolveTokenFeeAddress } from './configUtils.js';
 import { hypERC20contracts } from './contracts.js';
@@ -77,6 +78,7 @@ import {
   derivedIsmAddress,
   isEverclearTokenBridgeConfig,
   isMovableCollateralTokenConfig,
+  isXERC20TokenConfig,
 } from './types.js';
 
 type WarpRouteAddresses = HyperlaneAddresses<ProxyFactoryFactories> & {
@@ -158,6 +160,17 @@ export class EvmWarpModule extends HyperlaneModule<
     const actualConfig = await this.read();
     const transactions = [];
 
+    let xerc20Txs: AnnotatedEV5Transaction[] = [];
+    if (isXERC20TokenConfig(expectedConfig)) {
+      const { module, config } = await EvmXERC20Module.fromWarpRouteConfig(
+        this.multiProvider,
+        this.chainName,
+        expectedConfig,
+        this.args.addresses.deployedTokenRoute,
+      );
+      xerc20Txs = await module.update(config);
+    }
+
     /**
      * @remark
      * The order of operations matter
@@ -195,6 +208,8 @@ export class EvmWarpModule extends HyperlaneModule<
 
       ...this.createUpdateEverclearFeeParamsTxs(actualConfig, expectedConfig),
       ...this.createRemoveEverclearFeeParamsTxs(actualConfig, expectedConfig),
+
+      ...xerc20Txs,
 
       ...this.createOwnershipUpdateTxs(actualConfig, expectedConfig),
       ...proxyAdminUpdateTxs(
@@ -317,7 +332,6 @@ export class EvmWarpModule extends HyperlaneModule<
     actualConfig: DerivedTokenRouterConfig,
     expectedConfig: HypTokenRouterConfig,
   ): AnnotatedEV5Transaction[] {
-    actualConfig.type;
     if (
       !isMovableCollateralTokenConfig(expectedConfig) ||
       !isMovableCollateralTokenConfig(actualConfig)
@@ -359,7 +373,6 @@ export class EvmWarpModule extends HyperlaneModule<
     actualConfig: DerivedTokenRouterConfig,
     expectedConfig: HypTokenRouterConfig,
   ): AnnotatedEV5Transaction[] {
-    actualConfig.type;
     if (
       !isMovableCollateralTokenConfig(expectedConfig) ||
       !isMovableCollateralTokenConfig(actualConfig)