+++ typescript/cosmos-types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/cosmos-types
 
+## 25.0.0
+
 ## 24.0.0
 
 ## 23.0.0