+++ typescript/helloworld/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @hyperlane-xyz/helloworld
 
+## 25.0.0
+
+### Patch Changes
+
+- Updated dependencies [52ce778]
+- Updated dependencies [aaabbad]
+  - @hyperlane-xyz/utils@25.0.0
+  - @hyperlane-xyz/sdk@25.0.0
+  - @hyperlane-xyz/core@10.1.5
+
 ## 24.0.0
 
 ### Patch Changes