+++ typescript/rebalancer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hyperlane-xyz/rebalancer
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [52ce778]
+- Updated dependencies [aaabbad]
+  - @hyperlane-xyz/utils@25.0.0
+  - @hyperlane-xyz/sdk@25.0.0
+  - @hyperlane-xyz/core@10.1.5
+  - @hyperlane-xyz/metrics@0.1.3
+  - @hyperlane-xyz/provider-sdk@1.3.1
+
 ## 1.0.1
 
 ### Patch Changes